refactor(bigquery): lift IngestMode to shared common.utils.enums
The four ADBC bulk-ingest modes (create/append/replace/create_append) are
driver-agnostic, not BigQuery-specific. Move the IngestMode enum out of the
bigquery package into xorq.common.utils.enums (alongside ADBCBase) so every
ADBC backend can share and validate against it, and drop the now-empty
bigquery/enums.py.
